pipe api for platform for adding decrypted keys to any running ssh-agent detected. also did clang-tidy and cleaned up. downloaded llvm's run-clang-tidy.py script and modified it to only scan ascii-chat code (no dependencies) and use our config by defaults.. it's in scripts/. other random cleanup like removing debug logs and changing log statements from multiple to single.

+++ lib/image2ascii/ascii.c
@@ -499,7 +499,7 @@ char *ascii_create_grid(ascii_frame_source_t *sources, int source_count, int wid
       int dst_pos = dst_row * (width + 1) + h_padding;
       int copy_len = (line_len > width - h_padding) ? width - h_padding : line_len;
 
-      if (copy_len > 0 && (size_t)(dst_pos + copy_len) < target_size) {
+      if (copy_len > 0 && (size_t)dst_pos + (size_t)copy_len < target_size) {
         SAFE_MEMCPY(&result[dst_pos], target_size - (size_t)dst_pos, &src_data[line_start], (size_t)copy_len);
       }
